refactor(features): extract the feature/entitlement module — sequencing D5
src/lib/features/{entitlements,entitlements-service,global-features,
global-switches-service,registry}.ts -> src/modules/features/

5 files, 24 importers, 31 files rewritten across every reference form D1 taught
me to expect: the @/ alias, bare path literals in prose and config, escaped
regexes in test assertions, and relative imports.

THE OTHER HALF OF D5 ALREADY SHIPPED, and it was the important half. The plan's
D5 row reads "extract features; make both MCP servers read global_feature_switches;
retire the ACCESS_MODEL_V2 dual path". That last clause turned out to be a LIVE
SECURITY FAIL-OPEN rather than tidy-up — under v1 (the default, and prod's
config) a suspended caller's MCP write access was resolved by a second query
that degraded to not-suspended on any error. Retired and deployed to production
earlier today as SEC-83 (lyra-mcp-server#147), with the live build_sha verified
against main HEAD rather than inferred from a green smoke.

So this commit is the structural remainder: the module gets a name, and every
boundary rule about features becomes expressible.

Estate: modules.json (features -> src/modules/features/, owns.files re-derived),
env-access-baseline.json (entitlements-service keeps its single
PAID_LINKS_COMPLIANCE_READY read — the file moved, the read did not), the
test-path manifest, docs/ARCHITECTURE.md and kan416-derive.py.

Dated evidence deliberately untouched: the modularisation data snapshots, the F4
and F6 handovers, the dated plan, and the weekly run-ledger. Those are records of
what was true on a date; rewriting a path inside one falsifies the record rather
than fixing drift.

CTL-041 green (every tracked file under src/ still owned, owns.files matches the
tree), CTL-037 reports the baseline matches exactly, tsc clean.
