+++ src/app/dashboard/profile/actions.ts
@@ -6,7 +6,7 @@ import { sanitiseText, sanitiseUrl, type ActionResult } from '@/modules/guards/s
 import { moderateAndAudit } from '@/lib/moderation-audit';
 import type { WizardItem } from './steps/types';
 import { checkProfileWriteRateLimit } from '@/modules/guards/profile-rate-limit';
-import { getMyFeatureEntitlements } from '@/lib/features/entitlements';
+import { getMyFeatureEntitlements } from '@/modules/features/entitlements';
 import { isProviderAgeCheckActive, passedProviderAgeCheck, AGE_GATE_BLOCK_MESSAGE } from '@/lib/age/provider-gate';
 import { isAllowedProfileField } from './profile-fields';
 import { coerceVisibility } from './visibility';
