+++ src/modules/features/global-switches-service.ts
