+++ src/modules/features/registry.ts
