+++ tests/support/source-paths.json
@@ -1,7 +1,7 @@
 {
   "$comment": "GENERATED — regenerate with `npm run gen:test-paths`. Mirror of tests/support/source-paths.ts for .cjs and shell consumers (KAN-414 F4).",
   "generated_from": "literals referenced by tracked files under tests/",
-  "count": 191,
+  "count": 190,
   "SRC": {
     "codeowners": ".github/CODEOWNERS",
     "pullRequestTemplate": ".github/PULL_REQUEST_TEMPLATE.md",
@@ -160,7 +160,6 @@
     "postEvent": "src/lib/convene/post-event.ts",
     "resolveWidgets": "src/lib/dashboard/resolve-widgets.ts",
     "dbErrorCopy": "src/lib/db-error-copy.ts",
-    "entitlementsService": "src/lib/features/entitlements-service.ts",
     "inviteText": "src/lib/invite-text.ts",
     "moderationPolicy": "src/lib/moderation-policy.ts",
     "pipeline": "src/lib/recommender/v2/pipeline.ts",
