+++ tests/unit/media-uploads-gate.test.ts
@@ -33,7 +33,7 @@
  */
 import { uploadProfileFile } from '@/app/dashboard/profile/files-actions';
 import { uploadAvatar } from '@/app/dashboard/profile/actions';
-import { getMyFeatureEntitlements } from '@/lib/features/entitlements';
+import { getMyFeatureEntitlements } from '@/modules/features/entitlements';
 
 jest.mock('next/cache', () => ({ revalidatePath: jest.fn() }));
 
@@ -57,7 +57,7 @@ jest.mock('@/modules/platform/supabase-server', () => ({
   })),
 }));
 
-jest.mock('@/lib/features/entitlements', () => ({
+jest.mock('@/modules/features/entitlements', () => ({
   getMyFeatureEntitlements: jest.fn(),
 }));
